Kitchi injector fix

+++ code/modules/vore/fluffstuff/custom_items_vr.dm
@@ -1359,9 +1359,12 @@ obj/item/weapon/material/hatchet/tacknife/combatknife/fluff/katarina/handle_shie
 	icon = 'icons/obj/items.dmi'
 	icon_state = "dnainjector"
 
-/obj/item/weapon/fluff/kitchi_injector/attack_self(mob/user as mob)
+/obj/item/weapon/fluff/kitchi_injector/attack(mob/living/M, mob/living/user)
 
-	if(ishuman(user))
-		var/mob/living/carbon/human/H = user
-		H.monkeyize()
-	qdel(src) //One time use.
+	if(usr == M) //Is the person using it on theirself?
+		if(ishuman(M)) //If so, monkify them.
+			var/mob/living/carbon/human/H = user
+			H.monkeyize()
+			qdel(src) //One time use.
+	else //If not, do nothing.
+		to_chat(user,"<span class='warning'> You are unable to inject other people.</span>")
